+++ ktlint/src/main/kotlin/com/pinterest/ktlint/internal/FileUtils.kt
@@ -3,6 +3,7 @@ package com.pinterest.ktlint.internal
 import com.pinterest.ktlint.core.KtLint
 import com.pinterest.ktlint.core.LintError
 import com.pinterest.ktlint.core.RuleSet
+import com.pinterest.ktlint.core.api.FeatureInAlphaState
 import java.io.File
 import java.nio.file.FileSystem
 import java.nio.file.FileVisitResult
@@ -138,6 +139,7 @@ internal fun File.location(
 /**
  * Run lint over common kotlin file or kotlin script file.
  */
+@OptIn(FeatureInAlphaState::class)
 internal fun lintFile(
     fileName: String,
     fileContents: String,
@@ -148,7 +150,7 @@ internal fun lintFile(
     lintErrorCallback: (LintError) -> Unit = {}
 ) {
     KtLint.lint(
-        KtLint.Params(
+        KtLint.ExperimentalParams(
             fileName = fileName,
             text = fileContents,
             ruleSets = ruleSets,
@@ -167,6 +169,7 @@ internal fun lintFile(
 /**
  * Format a kotlin file or script file
  */
+@OptIn(FeatureInAlphaState::class)
 internal fun formatFile(
     fileName: String,
     fileContents: String,
@@ -177,7 +180,7 @@ internal fun formatFile(
     cb: (e: LintError, corrected: Boolean) -> Unit
 ): String =
     KtLint.format(
-        KtLint.Params(
+        KtLint.ExperimentalParams(
             fileName = fileName,
             text = fileContents,
             ruleSets = ruleSets,